backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
openai==1.52.0
anthropic==0.38.0
duckduckgo-search==7.3.0
reportlab==4.2.5
jinja2==3.1.4
pydantic==2.10.3
pydantic-settings==2.6.1